+++ 3team/Assets/Scripts/Navi/ARCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f741685a0beebc47a8050f66fae2e7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
