+++ 3team/Assets/Scripts/Navi/ARNavi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 44c4c2fa04bbc1a49b70941e373874f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
