+++ 3team/Assets/Scripts/Navi/Direction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a06ac5daa0057504098d9f67a2aba61e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
