+++ 3team/Assets/Scripts/Navi/MapProcessor1.cs
@@ -83,7 +83,7 @@ public class MapProcessor1 : MonoBehaviour
         while(true)
         {
             ProcessImage(Manager.UI.userPosition, buttonsPos);
-            yield return new WaitForSeconds(9f);
+            yield return new WaitForSeconds(3f);
         }
     }
     void ProcessImage(Vector2 userPosition, Vector2 buttonPosition)
@@ -96,8 +96,7 @@ public class MapProcessor1 : MonoBehaviour
 
             Vector2Int userPos = FindClosestRoadCoordinate(userPosition);
             Vector2Int buttonPos = FindClosestRoadCoordinate(buttonPosition);
-            Debug.Log(userPos);
-            Debug.Log(buttonPos);
+
             queue.Enqueue(userPos);
             parentMap[userPos] = userPos;
 
@@ -168,24 +167,5 @@ public class MapProcessor1 : MonoBehaviour
             closestY = Mathf.Clamp(closestY + 1, 0, gridSizeY - 1);
         }
         return new Vector2Int(closestX, closestY);
-    }
-
-    public void DisplayPath(List<Vector2Int> path)
-    {
-        // Line Renderer 컴포넌트를 가져옵니다.
-        LineRenderer lineRenderer = gameObject.GetComponent<LineRenderer>();
-
-        // Line Renderer를 초기화합니다.
-        lineRenderer.positionCount = path.Count;
-        lineRenderer.startWidth = 0.1f;
-        lineRenderer.endWidth = 0.1f;
-        lineRenderer.material = new Material(Shader.Find("Sprites/Default"));
-
-        // 경로의 각 지점을 Line Renderer에 추가합니다.
-        for (int i = 0; i < path.Count; i++)
-        {
-            Vector3 position = new Vector3(path[i].x, 0.1f, path[i].y); // 경로의 y값을 고정시켜 시각적으로 보이도록 합니다.
-            lineRenderer.SetPosition(i, position);
-        }
     }
 }
