+++ 3team/Assets/Scripts/Navi/NormalNavi.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c910ed7acaa8b0441b56c5285ae83a30
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
